Fixed Rendering Issue
Because of the way that the performance boost from the previous commit worked, the highlighted squares would not get removed from the RenderingServer.  The solution was to clear the RenderingServer RID during the engine's _draw() function, which executes asyncly, thus requiring the engine to redraw all text each time that the highlights need to be updated.

Overall, there is still a performance boost of roughly 10x faster execution compared to the original function, but slower than the previous commit.