Performance Boost
I was able to bring back the 80x performance boost.  When not resizing the matrix, rendering of the highlights is near instant.

NOTE: One time the app crashed with the message: "Stack Underflow (Engine Bug)"

This has only happened once and I have been unable to replicate the issue.  Hopefully it was a one-off bug.